# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Attaboy Platform API
  description: iGaming platform API — wallet ledger, sportsbook, payments, quests, social, and admin management.
  version: 1.0.0
  contact:
    name: Attaboy Platform
servers:
  - url: http://localhost:3100
    description: Local development
security: []
...
paths:
...
  /sportsbook/sports/{sportID}/events:
    get:
      tags: [Sportsbook]
      summary: List events for a sport
      operationId: listEvents
      security:
        - BearerAuth: []
      parameters:
        - name: sportID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Events list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SportsEvent"
        "400":
          $ref: "#/components/responses/ValidationError"
...
  /admin/sportsbook/events:
    get:
      tags: ["Admin: Sportsbook"]
      summary: List events (admin)
      operationId: adminListEvents
      security:
        - AdminAuth: []
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminEventSummary"
    post:
      tags: ["Admin: Sportsbook"]
      summary: Create event
      operationId: createEvent
      security:
        - AdminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventRequest"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
  /admin/sportsbook/events/{id}/status:
    patch:
      tags: ["Admin: Sportsbook"]
      summary: Update event status
      operationId: updateEventStatus
      security:
        - AdminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                score_home:
                  type: integer
                score_away:
                  type: integer
      responses:
        "200":
          $ref: "#/components/responses/StatusUpdated"
  /admin/sportsbook/events/{id}/settle:
    post:
      tags: ["Admin: Sportsbook"]
      summary: Settle all open bets for an event
      operationId: settleEvent
      security:
        - AdminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Settlement summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettleEventResult"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFoundError"

  # ── Admin: Reports ─────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    AdminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin JWT token
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    StatusUpdated:
      description: Status updated
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: updated
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
...
    SportsEvent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sport_id:
          type: string
          format: uuid
        league:
          type: string
        home_team:
          type: string
        away_team:
          type: string
        start_time:
          type: string
          format: date-time
        status:
          type: string
          enum: [upcoming, live, settled, cancelled]
        score_home:
          type: integer
        score_away:
          type: integer
...
    SettleEventResult:
      type: object
      properties:
        settled:
          type: integer
        won:
          type: integer
        lost:
          type: integer
        voided:
          type: integer
...
    CreateEventRequest:
      type: object
      required: [sport_id, league, home_team, away_team, start_time]
      properties:
        sport_id:
          type: string
          format: uuid
        league:
          type: string
        home_team:
          type: string
        away_team:
          type: string
        start_time:
          type: string
          format: date-time
    AdminEventSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sport_id:
          type: string
          format: uuid
        sport_name:
          type: string
        league:
          type: string
        home_team:
          type: string
        away_team:
          type: string
        start_time:
          type: string
          format: date-time
        status:
          type: string
        score_home:
          type: integer
        score_away:
          type: integer
...